Revise time elapse system into game time system.
The new approach simply tracks the current in-game time. The previous design assumed that we'd be keying lots of events off of a game-time clock. That, however, was a non-ES way of thinking about it. Now, the systems cue off of real time, and entire suites of systems can be toggled using SystemToggle.