Revise Task back to an int and introduce TaskFormatter for view handling.
By having user-facing Strings in the Task, we were definitely violating model-view separation. There really is no good reason to do that here. Also, 'Task' was the only component value that was still using a reference type, and it was doing so unnecessarily: the entity system in TriplePlay is optimized for reference types, so it's good practice for us to use it this way.