Split feature development into two entities.
One entity represents the development effort itself. It has a 'feature' component that links it to the other entity, which is the feature. The feature entity starts off disabled, and this way, we can give it usersPerSecond, for example, without that value actually being applied to the system. It makes sense from an ES POV: there is an entity for the feature, but it is not enabled until it is developed.