Refactor platform-specific date/time formatting.
The original implementation made use of a public static non-final field in ClockUtil, which is not very disciplined. The new approach includes subclasses of LwjglPlatform and HtmlPlatform, both of which implement SimGamePlatform. That interface specifies that implementors must provide a PlatformSpecificDateFormatter. This pattern could likely be used for any other platform-specific resources, should the need arise.

If we ever move to a different platform, the precondition in GameScreen should let us know if we forget to make a new custom platform implementation (as was done with SimGameHtmlPlatform and SimGameJavaPlatform).