Revise aspect ratio tool to compute bounding boxes, not layers and integrate with GameScreen.
This separation of concerns is important, since the layer behavior in PlayN/TriplePlay sometimes defies expectations. For example, the root layer of the whole game reports itself as having zero size throughout. The plan had been to use this as a spot to hold the resized active area, but that didn't work. GroupLayer sizes, generally, were troublesome. This approach allows us to use TPUI's absolute layout to specify a "content area" that has fixed bounds, as is shown now in GameScreen.