Add LearningSystem to increase workers' skills based on task
As part of this commit, world.developmentSkill was changed to a Component.FScalar so that it has the same type as world.maintenanceSkill. For both of these, using floats makes it easier to continuously add to a skill, yet for display (and computing impact), the float can be truncated to just its integer part.