Replace int task with Task interface
The new interface will allow the addition of custom tasks, whereas using integer constants was essentially creating an enumerated type. As part of this change, TaskFormatter was dropped, since Tasks now have a name() method---a more OO approach, but I think this will make it simpler to manage than somehow wedging this idea into an entity.

Also gone is the concept of "idle", except in unit tests. The CoreTask class now contains the two core tasks: DEVELOPMENT and MAINTENANCE.