Replace PreloadedImage with ImageCache
PreloadedImage was an enumerated type, which used to work with PlayN 1.x because of how it loaded assets from a static assets() source. In PlayN 2.0, however, it does not work this way: we have to inject the dependency on the Assets manually. The new ImageCache allows us to do this, which means we can write tests against ImageCache using mocks, which couldn't be done in enums because of how they are initialized at class loading time.