Use new font through stylesheet
This required a change in the implementation of the MainUIGroup.TaskSelector, since it was previously using Unicode to show a small triangle, but this font does not include that glyph. Now, I have used the polymer iron icon for arrow-drop-down. The icon also has to be manually toggled based on whether the selected task is reassignable or not.