Remove vulnerability system and status
This system had been updating the current exposure every update to be the sum of the vulnerabilities of completed features. This contradicted the design of the maintenance skill, which reduces exposure slowly over time. The solution, then, was to make the vulnerability of a feature add exactly once to company's exposure, just as the feature is completed; this behavior is captured in a unit test.