Convert game time to seconds
Game time had been expressed in milliseconds, but this led to integer overflows for longer periods such as 60 days (which is 5.184e+9 milliseconds). This did weird things when it overflowed, like swapping between two months. The solution here was to express game time in seconds, even though PlayN clock is in milliseconds.