Remove state pattern
The state pattern was introduced long ago as part of the original UI, which had a pause button. It hasn't been used since. This refactoring was simply pulling out dead code. The change touches GameWorld, since the pause system of GameScreen was the only thing referencing many of the systems by name; many systems are now anonymous--never referenced by name.