Refactor
Methods were using "advanceOneHour" as a proxy for the concept of "advance enough time that an exploit could have been generated", so this was pulled out into a method. However, there are still some tests based explicitly on the notion that exploits check each hour, so this may require more attention again later. (Once more, we are bit by the fact that we cannot simply use joda-time or something similar with explicit representation of periods, durations, etc).