Show only available workers in narrative event
Previously, workers were added to the event when the event is created, but this does not take into account that the worker may not be available when the event actually happens. For example, Esteban may be busy, and so he cannot be selected for some new task.

Now, the options are determined dynamically, so only non-busy workers are listed. That is, only those with reassignable tasks are listed.