Align intro skipping to CLI conventions
Previously the flag required arguments, but this violates the common conventions for command-line interaction: the undecorated command should do the expected thing, and a flag to toggle the behavior should not need an argument. Now, to skip the intro, one uses the flag "-skipIntro". If that flag is missing, the logo screen is played.

It strikes me that this term is currently overloading "intro" and "logo", and that may have to change later, but we'll deal with the problem we have, not the problem we may later have.

This solution also factors out the static field from LoadingScreen.