Revise platform-specific configuration
Previously, we were extending platform configuration to add new features. However, as I looked into Android, this is not possible, since the platform is created before the game is---we cannot swap in a new configuration as with the Java and HTML targets.

This design is actually a bit cleaner, now relying only on the fact that our game() is of type SimGame, rather than relying on game().plat also providing a specific interface.