Do not throw exception when re-initialized.
The Android backend is calling main() more than once, which is creating two SimGame instances, which calls initialize twice on the *Cache. I am not sure _why_ it is doing this, but it is, and the checkState was making the application crash. Unfortunately, we cannot easily turn this into a log warning either, because the logger is not initialized in unit tests environment (that is, SimGame.game will be null).