Add data breach event
This required changing some of the structures around narrative events. We needed to be able to generate text dynamically, and this is currently done with the NarrativeEvent.Text abstraction. I'm still not fully happy with it, because there's temporal coupling between text and run, and because we should have a way to run arbitrary code with an event, not just get its text.