Fix feature development scale
The tests were failing because the revised math in FeatureDevelopmentSystem broke assumptions in the unit test--namely, the assumption that the default scale for user skill was in development points per hour. To address this, I have introduced a company wide "inefficiencyFactor", which serves as a drag on development productivity. This produces the same effect as the previous change but without breaking the assumptions in the unit tests. (Are these assumptions valid? That much is hard to tell, but this was a convenient fix.)

The actual inefficiencyFactor used in the game is specified in the PlayableWorldFactory. Following this model should allow for much of the balancing logic to live in that one place.