Revise character UI
The character's full name is now shown on the main screen, not just their first name. The name is not shown in the profile box, nor are development and maintenance skills; this design decision was made in the spirit of not showing redundant information, although right now the connection between D/M and Development/Maintenance (and Feature/Exploits) is weak.