Revise narrative event structure
New narrative events are represented as subclasses of NarrativeEvent. If there's special processing to be done upon running, such as specifying how many users are lost or figuring out which workers are available, this can be done by overriding run() and calling super.run() at the end of the method.

The old approach was hard to manage: it required that new events specify data in the text() method before the option's run method is called, but that's really awkward. I am hopeful that this new approach will be more straightforward.