Revise image asset loading
The old approach used ImageCache, a singleton whose caching mechanism did not account for the time it takes to turn an image into a tile. This time is, I believe, the source of the error in which loaded images were not appearing on the LWJGL backend. The new GameAssets object, available in SimGame.game.assets, handles this properly, and so this fixes #3.