Prevent re-selecting current interaction area
This is implemented by disabling the button, although the current theme does not show the disabling as anything significant; a future change to the theme can fix this.

The reason for this change was to try to figure out why Frieda's face disappears when reselecting that view, but unfortunately this did not fix that.