Differentiate loading of images and tiles
Images that need to be in the background have to be prepared as tiles before they are used. However, there is a problem with Label's icons, where if a Tile is used, and the screen is reconfigured (revalidated?), the image disappears. The solution was to separate, at the level of asset loading, those things that are meant to be used as images (label icons) vs those to be used as tiles (backgrounds).