Show user trend line when data points are low
Previously, this was always a sparkline, but early in the game, this was flat and unintuitive. Now, when the number of data points are less than MINIMUM_X_UNITS, the line is shown advancing from the left. Only when it hits the right side (i.e. amount of data exceeds MINIMUM_X_UNITS) does it change to a sparkline visualization.