Revise exploit handling
An exploit now generates user loss for as long as it is not fixed,
which happens when its maintenanceProgress reaches 100%. Until then,
it has an attrition value that creates a loss in users per hour on
each update.

Note that exposure is not affected by this, since exposure is computed
based on the number of features, no the number of defects. I mention
this because I second-guessed myself a few times on this.