Replace viewSize dependency with playable screen bounds
Many parts of the code were computing percent of screen height using
SimGame.game.graphics().viewSize, but this did not take into account
that we are only using a smaller region of the available size---see
AspectRatioTool for how this works. This commit eliminates the
duplication of effort and distribution of this dependency, since now
all these calls go through a new GameBounds object at
SimGame.game.bounds.