Use text color alternation to indicate needs-attention
This was previously done with background blinking, which was too visually noisy. This approach is subtler, and the code was cleaned up a bit.

I tried to make this work with Coy's alternate icon, with the
exclamation point on it, but this had the problem that when the icon
was updated, the icon covered the text. Changing the highlight visual
design was much easier that hacking the button class.

This also reimplements the style class feature, which mysteriously
disappeared. Maybe it got killed in a failed merge.