Move default time scale to GameTimeSystem
This fixes #19 in a roundabout way. The problem reported was that changing the scale did not change how time passed, but the problem was we were changing the scale field rather than calling setGameTimeUnitsPerRealClockUnits---which was already being called in PlayableWorldFactory! Our change was being overwritten before it had a chance. Looking at the code, it feels like it makes more sense to have the default right in GameTimeSystem in case we go looking for it there again.