Use attention image backgrounds
This could not be done with Button icons because changing the icon
changed the z-ordering, putting the icon above the text. This approach
could also not be done with TriplePlay stylesheets, which require a
fixed background. By writing our own background, it can render
differently depending on the state of the button.