Restrict work to work hours
This introduces a new WorkHoursSystem that has workers working only
during traditional work hours. There is no special UI for this aside
from showing the "not at work" task when not at work.

This introduces a new notion for Tasks as well: tasks can be bound by
work hours. Such tasks can be stopped when a worker leaves and picked
back up when they return. Most tasks are like this, but not all, as is
shown by the Security Conference task, where a worker is gone for
continuous 48 hours.