Prevent boss from also being an employee.
I threw in a few tests to check around for this, although all the tests passed while the game still had the error. The problem deals with the order of evaluation of parameters---and wow, was it sloppy of me to write that code in the first place. I left the tests in because they don't hurt anything.

Part of introducing those tests was bringing in a new dependency: guava-testlib. This provides EqualsTester, which is a very nice way to test equals methods. This meant that the IML files had to change, too. While I was in there, I upgraded to guava's 19.0 release.