Enable firebase logging on HTML build
Networked logging is only enabled when the parameter "log=true" is
specified, as in the URL "127.0.0.1:8080/?log=true" used for local
testing. Sticking the parameter onto a URL we distribute will make
that version log as well.

Only INFO-level log messages are recorded to Firebase, and right now,
there is only one of these: the message that the game has loaded
specified in SimGameHtml. Although more logging information needs to
be added in order for these data to be useful, it can still be said
that this fixes #30, since it provides all the necessary
infrastructure.

Each log message has a timestamp, specified in milliseconds since the
Unix epoch. Although it's not human-readable, it will be easier to
analyze later than if we used human-readable timestamps here.