Ensure CEO cannot be a worker
This fixes #38. The problem had to do with EmployeePool, and that it was creating a new list of available employees from the master list each time recruit was called. There are some assorted tests around now to catch some of these cases.

As part of the triangulation process, I also added some equals and hashCode methods to immutable model-level classes.