Start new game logs for each game
Previously all the game logs from the same session were going to the same log in Firebase. Now, there is a separate list of events for each game instance. To get this work, the collector watches for a very specific log message---LogUtils.START_GAME_MESSAGE---which triggers it to make a new entry on Firebase.

This fixes #40