Show post-game survey when consent has been given
Part of this commit includes removing the multiple playn.scene.Pointer
instantiations, which were both unnecessary and causing problems when
screens were overlapping.

The text on the EndScreen button was changed because no longer do we go right back to the start screen: if consent was given, we go to the survey screen.