Prevent popups when the button is disabled
There are some strange cases where we get notification of a button click, but the button is disabled. I expect it's a race condition between the event-routing and the system that turns off the buttons. The problem arose on the Java target at least, but the guard here prevents it.