Use icons for change view button rather than custom backgrounds
The custom backgrounds were, I believe, the source of a graphics memory leak. The repeated calls to `canvas.toTexture()` look like they were generating new texture instances each time, not just getting a texture version of the canvas.

This commit is missing some important features, like showing any kind of "needs attention" indication and the number of features & exploits, but it's a stake in the ground.