Make employees expand/contract properly on mobile
This fixes #62

Note that the Java and HTML platforms have different ways of expressing when a pointer event has finished, and so there's platform-specific code to deal with touch events. There's still a problem, though, that Chrome handles these events differently from Firefox, and so a different generalizable solution may be required.